[jsk_recognition_utils] Separate line.h from geo_util.h
