+++ jsk_recognition_utils/src/geo_util.cpp
@@ -66,165 +66,7 @@ namespace jsk_recognition_utils
     rot.col(2) = ez.normalized();
     return Eigen::Quaternionf(rot);
   }
-  
-  Line::Line(const Eigen::Vector3f& direction, const Eigen::Vector3f& origin)
-    : direction_ (direction.normalized()), origin_(origin)
-  {
-
-  }
-
-  void Line::getDirection(Eigen::Vector3f& output) const
-  {
-    output = direction_;
-  }
-
-  Eigen::Vector3f Line::getDirection() const
-  {
-    return direction_;
-  }
-  
-  void Line::getOrigin(Eigen::Vector3f& output) const
-  {
-    output = origin_;
-  }
-
-  void Line::foot(const Eigen::Vector3f& point, Eigen::Vector3f& output) const
-  {
-    const double alpha = computeAlpha(point);
-    output = alpha * direction_ + origin_;
-  }
-
-  double Line::distanceToPoint(
-    const Eigen::Vector3f& from, Eigen::Vector3f& foot_point) const
-  {
-    foot(from, foot_point);
-    return (from - foot_point).norm();
-  }
-  
-  double Line::distanceToPoint(const Eigen::Vector3f& from) const
-  {
-    Eigen::Vector3f foot_point;
-    return distanceToPoint(from, foot_point);
-  }
-
-  double Line::angle(const Line& other) const
-  {
-    double dot = fabs(direction_.dot(other.direction_));
-    if (dot > 1.0) {
-      return M_PI / 2.0;
-    }
-    else {
-      double theta = acos(dot);
-      if (theta > M_PI / 2.0) {
-        return M_PI / 2.0 - theta;
-      }
-      else {
-        return theta;
-      }
-    }
-  }
-
-  bool Line::isParallel(const Line& other, double angle_threshold) const
-  {
-    return angle(other) < angle_threshold;
-  }
-
-  bool Line::isPerpendicular(const Line& other, double angle_threshold) const
-  {
-    return (M_PI / 2.0 - angle(other)) < angle_threshold;
-  }
-
-  bool Line::isSameDirection(const Line& other) const
-  {
-    return direction_.dot(other.direction_) > 0;
-  }
-
-  Line::Ptr Line::flip()
-  {
-    Line::Ptr ret (new Line(-direction_, origin_));
-    return ret;
-  }
-  
-  Line::Ptr Line::midLine(const Line& other) const
-  {
-    Eigen::Vector3f new_directin = (direction_ + other.direction_).normalized();
-    Eigen::Vector3f new_origin;
-    other.foot(origin_, new_origin);
-    Line::Ptr ret (new Line(new_directin, (new_origin + origin_) / 2.0));
-    return ret;
-  }
-
-  void Line::parallelLineNormal(const Line& other, Eigen::Vector3f& output)
-    const
-  {
-    Eigen::Vector3f foot_point;
-    other.foot(origin_, foot_point);
-    output = origin_ - foot_point;
-  }
-  
-  Line::Ptr Line::fromCoefficients(const std::vector<float>& coefficients)
-  {
-    Eigen::Vector3f p(coefficients[0],
-                      coefficients[1],
-                      coefficients[2]);
-    Eigen::Vector3f d(coefficients[3],
-                      coefficients[4],
-                      coefficients[5]);
-    Line::Ptr ret(new Line(d, p));
-    return ret;
-  }
-
-  double Line::distance(const Line& other) const
-  {
-    Eigen::Vector3f v12 = (other.origin_ - origin_);
-    Eigen::Vector3f n = direction_.cross(other.direction_);
-    return fabs(n.dot(v12)) / n.norm();
-  }
-
-  Line::Ptr Line::parallelLineOnAPoint(const Eigen::Vector3f& p) const
-  {
-    Line::Ptr ret (new Line(direction_, p));
-    return ret;
-  }
-  
-  double Line::computeAlpha(const Point& p) const
-  {
-    return p.dot(direction_) - origin_.dot(direction_);
-  }
-  
-  PointPair Line::findEndPoints(const Vertices& points) const
-  {
-    double min_alpha = DBL_MAX;
-    double max_alpha = - DBL_MAX;
-    Point min_alpha_point, max_alpha_point;
-    for (size_t i = 0; i < points.size(); i++) {
-      Point p = points[i];
-      double alpha = computeAlpha(p);
-      if (alpha > max_alpha) {
-        max_alpha_point = p;
-        max_alpha = alpha;
-      }
-      if (alpha < min_alpha) {
-        min_alpha_point = p;
-        min_alpha = alpha;
-      }
-    }
-    // JSK_ROS_INFO("min: %f", min_alpha);
-    // JSK_ROS_INFO("max: %f", max_alpha);
-    return boost::make_tuple<Point, Point>(min_alpha_point, max_alpha_point);
-  }
-
-  void Line::print()
-  {
-    JSK_ROS_INFO("d: [%f, %f, %f], p: [%f, %f, %f]", direction_[0], direction_[1], direction_[2],
-             origin_[0], origin_[1], origin_[2]);
-  }
-
-  void Line::point(double alpha, Eigen::Vector3f& output)
-  {
-    output = alpha * direction_ + origin_;
-  }
-  
+    
   Segment::Segment(const Eigen::Vector3f& from, const Eigen::Vector3f to):
     Line(from - to, from), from_(from), to_(to)
   {
